Adds windows support. To facilitate this:

- Switched from GLEW to GLAD.
- Switched from using system GLFW and GLM to building in-tree.
- Removed Boost. Replaced Boost program_options with cxxopts.
- Fixed use-after-free in gui::file_dialog::render.
- Destruct app object before terminating GLFW.
- Added Windows build instructions.

+++ src/formats/vif.h
@@ -139,6 +139,6 @@ struct vif_packet {
 
 std::vector<vif_packet> parse_vif_chain(const stream* src, std::size_t base_address, std::size_t qwc);
 
-uint64_t bit_range(uint64_t val, int lo, int hi);
+int bit_range(uint64_t val, int lo, int hi);
 
 #endif
